requirements.txt
Django==1.6.5
Fabric==1.9.0
South==1.0
Werkzeug==0.9.6
argparse==1.2.1
django-appconf==0.6
django-backupdb==0.5.5
django-compressor==1.4
django-debug-toolbar==0.11
django-dotenv==1.2
django-extensions==1.3.8
django-pyscss==1.0.1
ecdsa==0.11
fusionbox-fabric-helpers==0.5.0
paramiko==1.14.0
psycopg2==2.5.3
pyScss==1.2.0.post3
pycrypto==2.6.1
python-memcached==1.53
raven==5.0.0
six==1.7.3
sqlparse==0.1.11
termcolor==1.1.0
wsgiref==0.1.2
